+++ framework/core/src/Events/DiscussionWasHidden.php
@@ -1,0 +1,31 @@
+<?php
+/*
+ * This file is part of Flarum.
+ *
+ * (c) Toby Zerner <toby.zerner@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Flarum\Events;
+
+use Flarum\Core\Discussions\Discussion;
+
+class DiscussionWasHidden
+{
+    /**
+     * The discussion that was hidden.
+     *
+     * @var Discussion
+     */
+    public $discussion;
+
+    /**
+     * @param Discussion $discussion
+     */
+    public function __construct(Discussion $discussion)
+    {
+        $this->discussion = $discussion;
+    }
+}